Added multithreading support, switched from floats to doubles for increased precision.

+++ Include/Canvas.h
@@ -8,9 +8,9 @@ namespace photon {
         
         std::uint32_t width;
         std::uint32_t height;
-        float distance;
+        double distance;
 
-        Canvas(std::uint32_t width, std::uint32_t height, float distance) :
+        Canvas(std::uint32_t width, std::uint32_t height, double distance) :
             width(width), height(height), distance(distance) {}
 
     };
