+++ Include/Intersection.h
@@ -6,10 +6,10 @@ namespace photon {
 
     struct Intersection {
 
-        float t;
+        double t;
         Vec3f normal;
 
-        Intersection(float t, const Vec3f& normal)
+        Intersection(double t, const Vec3f& normal)
             : t(t), normal(normal) {}
 
     };
