+++ Source/MathUtils.cpp
@@ -6,11 +6,11 @@ namespace photon {
 
     Vec2f::Vec2f() : x(0.0f), y(0.0f) {}
 
-    Vec2f::Vec2f(float x, float y) : x(x), y(y) {}
+    Vec2f::Vec2f(double x, double y) : x(x), y(y) {}
 
     Vec3f::Vec3f() : x(0.0f), y(0.0f), z(0.0f) {}
 
-    Vec3f::Vec3f(float x, float y, float z) : x(x), y(y), z(z) {}
+    Vec3f::Vec3f(double x, double y, double z) : x(x), y(y), z(z) {}
 
     Vec3f Vec3f::operator+(const Vec3f& other) const {
         return Vec3f(x + other.x, y + other.y, z + other.z);
@@ -24,7 +24,7 @@ namespace photon {
         return Vec3f(x - other.x, y - other.y, z - other.z);
     }
 
-    Vec3f Vec3f::operator*(float s) const {
+    Vec3f Vec3f::operator*(double s) const {
         return Vec3f(x * s, y * s, z * s);
     }
 
@@ -39,8 +39,8 @@ namespace photon {
         return *this;
     }
 
-    float Vec3f::length() const {
-        return sqrtf(x * x + y * y + z * z);
+    double Vec3f::length() const {
+        return sqrt(x * x + y * y + z * z);
     }
 
     Vec3f Vec3f::normalize() const {
@@ -48,11 +48,11 @@ namespace photon {
         return Vec3f(x / len, y / len, z / len);
     }
 
-    float Vec3f::distance(const Vec3f& other) const {
+    double Vec3f::distance(const Vec3f& other) const {
         return (other - *this).length();
     }
 
-    float Vec3f::dot(const Vec3f& other) const {
+    double Vec3f::dot(const Vec3f& other) const {
         return x * other.x + y * other.y + z * other.z;
     }
 
@@ -67,7 +67,7 @@ namespace photon {
     Ray::Ray(const Vec3f& origin, const Vec3f& direction)
         : origin(origin), direction(direction.normalize()) {}
 
-    Vec3f Ray::point_at(float t) const {
+    Vec3f Ray::point_at(double t) const {
         return origin + direction * t;
     }
